Moved post-test VMMC update function to circumcision module.
